# Make RSA_check_key more than 2x as fast.